pyproject.toml
[tool.poetry]
name = "cyberos"
version = "0.1.0"
description = "AI-NATIVE OS"
authors = ["ZhouYimao <731901968@sjtu.edu.cn>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain-openai = "^0.1.20"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
langgraph = "^0.2.3"
langgraph-checkpoint = "^1.0.2"
langgraph-checkpoint-postgres = "^1.0.3"
langgraph-checkpoint-sqlite = "^1.0.0"
langchain-community = "^0.2.11"
llama-index = "^0.10.64"
llama-index-graph-stores-neo4j = "^0.2.14"
llama-index-embeddings-langchain = "^0.1.2"
llama-index-llms-langchain = "^0.3.0"
selenium = "^4.23.1"
pymupdf = "^1.24.9"
unstructured = {extras = ["all-docs"], version = "^0.15.7"}
libmagic = "^1.0"
pdfminer = "^20191125"
pdfminer-six = "^20240706"
pillow-heif = "^0.18.0"
matplotlib = "^3.9.2"
pytesseract = "^0.3.13"
unstructured-pytesseract = "^0.3.13"
llmsherpa = "^0.1.4"
python-docx = "^1.1.2"
docx2txt = "^0.8"
rapidocr-onnxruntime = "^1.3.24"
aiohttp = "^3.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
